Fix warnings
git-svn-id: http://svn.macosforge.org/repository/mdnsresponder/trunk@539 214c2c4a-bf3b-4dcf-9390-e4dd3010487d

+++ mDNSMacOSX/CFSocket.c
@@ -229,7 +229,7 @@ mDNSlocal void myCFSocketCallBack(CFSocketRef s, CFSocketCallBackType type, CFDa
 		verbosedebugf("myCFSocketCallBack got a packet from %.4a to %.4a on interface %.4a/%s",
 			&senderaddr, &destaddr, &info->ifinfo.ip, info->ifa_name);
 
-	if (err < sizeof(DNSMessageHeader)) { debugf("myCFSocketCallBack packet length (%d) too short", err); return; }
+	if (err < (int)sizeof(DNSMessageHeader)) { debugf("myCFSocketCallBack packet length (%d) too short", err); return; }
 	
 #if mDNS_AllowPort53
 	if (s == info->cfsocket53)
